+++ memoriter-website-gatsby/src/components/blog/blog-post-layout.tsx
@@ -1,8 +1,8 @@
 import React, { FC } from 'react';
 import * as styles from '../../styles/blog/blog-post-main.module.scss';
-import WebsiteWrapper from '../wrapper/website-wrapper';
-import BlogPostHeader from '../blog/blog-post-header';
-import BlogPostFooter from '../blog/blog-post-footer';
+import WebsiteLayout from '../layout/website-layout';
+import BlogPostHeader from './blog-post-header';
+import BlogPostFooter from './blog-post-footer';
 
 interface Props { // type definitions for props
   topic: string;
@@ -20,7 +20,7 @@ const BlogPost: FC<Props> = ({ topic, date, author, title, linkedBlogs, wordCoun
   const minutesRead = (wordCount / 250).toFixed() // (slightly inaccurate) number of words, 250 is an estimation for average words read per minute
 
   return (
-    <WebsiteWrapper currentPage='blog'>
+    <WebsiteLayout currentPage='blog'>
 
       <section className={styles.blog_post_main}>
 
@@ -35,7 +35,7 @@ const BlogPost: FC<Props> = ({ topic, date, author, title, linkedBlogs, wordCoun
       {/*footer with the read more links, share options etc.*/}
       <BlogPostFooter title={title} linkedBlogs={linkedBlogs} allBlogPosts={allBlogPosts}/>
 
-    </WebsiteWrapper>
+    </WebsiteLayout>
   );
 };
 export default BlogPost;
