+++ memoriter-website-gatsby/src/components/layout/cookie-banner/cookie-banner.tsx
