+++ memoriter-website-gatsby/src/components/layout/window-size-alert.tsx
