+++ memoriter-website-gatsby/src/pages/download.tsx
@@ -1,14 +1,14 @@
 import React, { FC } from 'react';
 import { HeadFC } from 'gatsby';
-import WebsiteHead from '../components/website-head';
-import WebsiteWrapper from '../components/wrapper/website-wrapper';
+import WebsiteHead from '../components/layout/website-head';
+import WebsiteLayout from '../components/layout/website-layout';
 import DownloadBanner from '../components/download/download-banner';
 import DownloadMain from '../components/download/download-main';
 import DownloadMobile from '../components/download/download-mobile';
 
 const Download: FC = () => {
     return (
-        <WebsiteWrapper currentPage='download'>
+        <WebsiteLayout currentPage='download'>
             {/*banner with download buttons*/}
             <DownloadBanner/>
 
@@ -18,7 +18,7 @@ const Download: FC = () => {
             {/*mobile app download*/}
             <DownloadMobile/>
 
-        </WebsiteWrapper>
+        </WebsiteLayout>
     );
 }
