+++ memoriter-website-gatsby/src/pages/impressum.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import { HeadFC } from 'gatsby';
 import '../styles/legal.scss';
-import WebsiteHead from '../components/website-head';
-import WebsiteWrapper from '../components/wrapper/website-wrapper';
+import WebsiteHead from '../components/layout/website-head';
+import WebsiteLayout from '../components/layout/website-layout';
 
 const Impressum: FC = () => {
     return (
-        <WebsiteWrapper currentPage=''>
+        <WebsiteLayout currentPage=''>
 
             {/*main body with text*/}
             <article className='legal-main'>
@@ -31,7 +31,7 @@ const Impressum: FC = () => {
                 
             </article>
             
-        </WebsiteWrapper>
+        </WebsiteLayout>
     );
 }
