+++ memoriter-website-gatsby/src/pages/terms.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import { HeadFC } from 'gatsby';
 import '../styles/legal.scss';
-import WebsiteHead from '../components/website-head';
-import WebsiteWrapper from '../components/wrapper/website-wrapper';
+import WebsiteHead from '../components/layout/website-head';
+import WebsiteLayout from '../components/layout/website-layout';
 
 const Terms: FC = () => {
     return (
-        <WebsiteWrapper currentPage=''>
+        <WebsiteLayout currentPage=''>
 
             {/*main body with text*/}
             <article className='legal-main'>
@@ -21,7 +21,7 @@ const Terms: FC = () => {
 
             </article>
             
-        </WebsiteWrapper>
+        </WebsiteLayout>
     );
 }
